posix: fix O_DIRECT usage
On Linux, O_DIRECT requires the buffer to be aligned at logical sector
size boundary, and both buffer size and file offset should be multiple
of the logical sector size as well. Usually it is equal to 512 bytes and
may be obtained by BLKBSZGET ioctl() on a device special file. On the
other side O_DIRECT may give better performance with chunks aligned/sized
to multiple of the logical filesystem block size. The latter is obtained
via statvfs() in posix_init() and used to check whether the chunk is
suitable for O_DIRECT where applicable.

Also fix false positive tests/bugs/glusterfs/bug-866459.t to check
whether the test file is not truncated.

Signed-off-by: Dmitry Antipov <dantipov@cloudlinux.com>
Updates: #1000