snapshots: Save Snapshot Volume ID and Snapshot name in the restored volume
During Glusterd restart, it tries to mount the Restored
Volume/Clone/Snapshot if it is in started state or if any
Snapshot is in activated state.

Restored Volume fails to find the Snapshot name from which it was
restored. This PR adds snapshot volume_id and snapshot name to
the volume info of restored volume.

Change-Id: I198c854e21297ae0353de49cafff049726f8eb03
Signed-off-by: Aravinda Vishwanathapura <aravinda@kadalu.io>