improve judgment standard for gf_is_service_running
We can not judge simplely if the service is running via existence of
pid. If a gluster service exited abnormally, the pid file may remain.
If the pid is assigned to another new process, it will be killed.
So the judgment criteria should be more rigorous. The precondition is
the pid file can not be locked.

Fixes: #2995
Signed-off-by: Yun Zhou <yun.zhou@windriver.com>