qr_lookup()/qr_lookup_cbk() - remove unneeded dict_get()
There was a relic from the past that tried to fetch a no longer relevant
'sh-failed' dict entry - removed it, and cleaned up the code a bit around.

TODO (in a different patch): store the result of
data_from_uint64(conf->max_file_size)
in the conf and fetch it, instead of re-doing it on every lookup.

Fixes: #2975
Signed-off-by: Yaniv Kaul <ykaul@redhat.com>