glusterd: Update cutom log directory option (#2957)
**Description:**
Currently option `cluster-test-mode` was used to set
a user-defined logging directory for gluster related logs which
doesn't seem to be an appropriate name for the option.

**Fix:**
Updated the option to be `logging-directory` keeping in mind the
naming convention of other options like `working-directory`,
`run-directory`.

**NOTE:**
This option doesn't updates the path for cli and glusterd log file,
that still needs to be set manually via command line or through the
sysconfig file.

Updates: #2939

> Updates: #2939

> Change-Id: I5fbbeff21ea1a89d439537311a81389b57e7acde
> Signed-off-by: nik-redhat <nladha@redhat.com>

Change-Id: I137b0aa529f17c6bf95018db6ad28fbfd8681e75
Signed-off-by: nik-redhat <nladha@redhat.com>