configure: Enable ltcmalloc_minimal instead of ltcmalloc
The gluster process is getting crashed while ltcmalloc
library is loaded/unloaled in two different threads.The process
is getting crashed during access of tls variables in heap profiler
api. Though the issue is already fixed(https://github.com/gperftools/gperftools/issues/786)
in gperftool-9 but the rpm has not been build for rhel-(7|8).The heap profile api is access
only an application is linked with libtcamlloc.

Solution: To avoid a crash linked library with ltcmalloc_minimal
Fixes: #2971
Signed-off-by: Mohit Agrawal <moagrawa@redhat.com>