test: bug-1319374-THIS-crash.t is getting crashed
The test case (./tests/bugs/gfapi/bug-1319374-THIS-crash.t) is
continuously getting crashed on the pull request (#2940).
The test case is crashed because openssl does not allow
to call SSL_library_init multiple times in the
multi-threading programs and the test program is trying to
call SSL_library_init more than once so it is crashing.

Solution: Call SSL_Library_init by gfapi library instead of
          calling by socket.so to avoid a crash.

Fixes: #3026
Signed-off-by: Mohit Agrawal <moagrawa@redhat.com>