posix_disk_space_check, dht_disk_usage: save multiplcation if using percentage
If we are using percentage of free space, there's no need to multiply
both the free space and the and the total size by buf.f_bsize,
as we compare one to each other.

Similar work in DHT's code.

Also, switched from using a char ('p') to signify we are using percentage
for disk usage to a simple boolean. True means we are using percentage.

Updates: #1000
Signed-off-by: Yaniv Kaul <ykaul@redhat.com>