snapshot: Illegal access coverity issue
**Issue**
The check for volinfo added earlier didn't solved the issue completely.
The glusterd_snapshot_revert_partial_restored_vol() frees the volinfo
in the end but no where assigns NULL to it. Hence, as a result
`if(!volinfo)` will still be false but the pointer points to a
freed memory and dereferencing it would result in a crash.

**Fix**
We need to unmount the snap after the restore but if the volinfo is
deleted in the above function (glusterd_snapshot_revert_partial_restored_vol)
then it won't be done. Hence, moved the `glusterd_volinfo_unref()` to
the caller function.

CID: 1395617
Updates: #1060

Change-Id: I406df209189971eacd1d99bda575ddb1c09af89d
Signed-off-by: nik-redhat <nladha@redhat.com>