snapview-server: fix a possible extra 'unref()' on snapshot root inode
`table->root` inode is agnostic to ref/unref as per the current inode
table implementation, but when in case of `snapd` process, the root inode
of snapshot process is mapped to another directory inode in global inode
table. Hence we need some 'extra' protection for this 'root' inode from
the snapshot process. Add an extra ref to the inode which goes through
glfs object initialization. This would prevent a possible mismanagement
of root inode during de-activate of snapshot.

Updates: #3103
Change-Id: I12b9c85f677c2868ef112f36547eb69dc80d3b7b
Signed-off-by: Amar Tumballi <amar@kadalu.io>