fix: Improved macOS version assertion in configure.
On major versions of macOS > 10, this assertion could fail if the minor
version of the developer's OS is less than 7. This corrects that by also
assertion that the major version can be greater than 10. When the
version is greater than 10 the minor version is no longer important for
this assertion.

Resolves #3166