Introduce '--fuse-setlk-handle-interrupt' option
TL;DR this is a workaround to certain issues with
locking.

In detail:

A number of semantic issues came up in relation to
locking, due to which the jolly optimistic approach
to lock interruption -- "well, we just make use of
whatever cancellation primitives the locking API has
to offer" -- is not always a viable choice.

Solidifying the lock subsystem is an ongoing effort;
however, until this reaches completion, we need a
workaround, which, in this case, boils down to
allowing users (mounters) to disable interruptability
of locks via a gliusterfs command line / mount option.

The issues that most obviously interfere with lock
interruption are: #3177, #3179. (Note that by now
the former one is resolved and closed; I'm mentioning
it now as the referred misbehavior might still be present
on older Glusterfs versions where this commit might get
backported.)

As there is no easy way to fix some of the referred
issues, the default valie of the option shall be 'off'.

Update: #3179
Change-Id: I28568d6182b2e716b36995ee28b26726af633942
Signed-off-by: Csaba Henk <csaba@redhat.com>