[configure] Improve sanitizer detection & linking
It is not a good idea to check for symbol `__xsan_init` as its newer
alternative `__xsan_init_vN` changes everytime the ABI changes.
`_init` symbol is present in all of the sanitizers as is also backwards
compatible.

Older versions of `gcc` were unable to link properly with sanitizers but
as of ver 5 they can link properly using `-fsanitize=xxx`. Using
`-lxsan` is discouraged as it misses a few extra flags that `-fsanitize`
includes. Bear in mind that flags `-W1,-z,defs` will cause linker to
fail as san runtime is not linked when linking shared libraries.

Change-Id: I12e0a431feabfc9cee22503ed5afa7e06b44c366
Signed-off-by: black-dragon74 <niryadav@redhat.com>