Arbiter: Do not compare file size against arbiter brick
Problem:
In afr_selfheal_unlocked_inspect(), if there are no pending xattrs
set for the entries inside indices folder it checks for other fileds
to identify anything is actually pending. Doing so, it will check for
the file size on all the bricks. If the volume is of arbiter type, the
size comparision will always fail for the files which does have some
data in it, when compared against the arbiter brick.

Fix:
Do not check for the file size against arbiter brick.

Fixes: #3200
Change-Id: I9ed31a49ae9fcb1a99dec2933a66e27af65fbf78
Signed-off-by: karthik-us <ksubrahm@redhat.com>