[autogen] Add check for pkg-config
On some platforms, pkg-config is not bundled with aclocal or is installed
at some non standard location due to which aclocal is unable to find it. It leads
to some very unhelpful error messages when running configure. This patch adds a check
for the same. See https://tirania.org/blog/archive/2012/Oct-20.html for more on this.

Signed-off-by: black-dragon74 <niryadav@redhat.com>