afr/upcall: Skip upcall notification from Arbiter
We never used to select Arbiter node as read subvol, hence
it is not a wise decesion to send an upcall notification
from Arbiter node that has an iatt update flag.

But we may need to invalidate the entries if pending xatrrs
are present on the dict. So we strip out only the flag that
request iatt update.

Change-Id: I3629d33019c93944b61b0de2d040000cb95a64df
Signed-off-by: Mohammed Rafi KC <rafi.kavungal@iternity.com>