afr/notify: Fix notify argument for upcall
upcall notification expects the upcall data struct
gf_upcall_cache_invalidation as the main argument.
While AFR requires another argument to identify the
child xlator which send the upcall. As of now, afr_notify
was using the upcall strcutre as the xlator variable
and was calculating the child xlator index value to
a wrong value.

This patch fixes the child index calculation, as well
as tries to have a unifrom way to calculate it. This
patch has also moved the hndling of notification for
GF_EVENT_TRANSLATOR_OP as the first one to process since
this event doesnot have any meaningful child_xlator
computation.

Change-Id: I7bf03aea53fdedec298869a82efd9b2c6b4e867c
Signed-off-by: Mohammed Rafi KC <rafi.kavungal@iternity.com>