afr/upcall: Optimize upcall event handling
If the pending keys are present on the dict, it might not
be a good idea to blindly invalidate the cache. We can go
one step further and check the pending key value, Since
it is also possible that an already healed file can have
pending key set with zero values.

Change-Id: I29bd4fd688649b5aa151321e8a043729032ad0dd
Signed-off-by: Mohammed Rafi KC <rafi.kavungal@iternity.com>