nfs: Optimize ctxcount value to reduce memory consumption for nfs-server (#3237)
In case of gnfs for nfs-server all volumes are attached with a single graph.
During inode table creation it sets ctxcount same to the total xlators are
attached with a graph. If a user has created 100 volumes (1x3) so total
number of xlators are attached with graph is 702 per volumes almost 7
xlators are attached (io-stat, dht, write-behind, afr, c1, c2, c3) so total
ctxcount set is 702. During inode creation we do allocate memory based on
number of ctxcount to save the xlator data on inode. In case of 100 volumes
memory allocation per inode will be around 32(bytes per ctxcount) * 702 = 22464 Bytes
for ctx.In caseof gnfs default lru_limit is 90000 so total memory allocation will
be 22464*90000(almost 2G). Per volume memory consumption is 2G so it will be huge
for all 100 volumes if every volumes has reached till maximum lru_limit.

Solution: An inode is associated with a specific volume so it is not
          possible multiple volume can save the data on inode_ctx so
          set the ctxcount based on the xlators associated with a volume
          not all the volumes.With this change It will reduce huge
          memory consumption for nfs-server. After set ctxcount (702/100 + 2 = 9)
          total memory allocation will be 32 * 9 * 90000 = 26M per volumes
          it is a huge saving and even for all 100 volumes has reached at
          maximum lru_limit memory consumption will not be more than 2.6G.

Change-Id: If35804d4337723ea902a471ca28842a595030e78
Fixes: #3236
Signed-off-by: Mohit Agrawal <moagrawa@redhat.com>