upcall: upcall_cache_invalidate() should not check if upcall is enabled
It was already called only if upcall was enabled, in all places but one.
Added that one (up_writev_cbk) as well and removed the check from
upcall_cache_invalidate() which was useful in case where it is called in a loop,
for example in up_readdirp_cbk()

In addition, wherever it was called with a constant flag, used that constant.

Updates: #1000
Signed-off-by: Yaniv Kaul <ykaul@redhat.com>