event-epoll: move slots_used out of the slots and into a table structure
For some reason, every slot had a 'slots_used' which was wasteful, as this is
a per-table value, not a per slot. Re-arranged the code to have a table structure
with a table_slots_used variable to store how many slots used.

(In the future, I think it makes sense to move it to an 64bit bitmap and have 64 entries per table, instead of 1024)

While at it, re-arranged __event_slot_alloc() code a bit.

Updates: #3228
Signed-off-by: Yaniv Kaul <ykaul@redhat.com>