server:Optimize server_process_event_upcall code path
The function server_process_event_upcall handles upcall event and send a
notification to client. To send a notification to client it has taken
a lock to access xprt and the lock has not released until the event has not
been sent to the client, It means at a time it will not able to process
more than one notification it has lock contention while a brick process is trying
to send more than one upcall event notification.

Solution: Instead of sending an event under critical section take a reference
          on the xprt.
Fixes: #3321
Signed-off-by: Mohit Agrawal <moagrawa@redhat.com>

Change-Id: Ibaab1626ab6e0c318df77e8efe528e71f3802517