meta.c/.h: meta_lookup() small improvements
The meta_lookup() function would use two macros:
META_HOOK - which would call THIS to fetch this->private
IS_META_ROOT_GFID - which did a strcmp() to the result of a uuid_utoa() function.

Simplified by saving the result of the uuid_utoa() in the private struct, doing the comparison directly, and without the call to 'THIS'.
In turn, re-used that result to copy that UUID instead of parsing it (again).

Found out you can skip most of this if 'loc->name' is null, so added it as well.

Updates: #3359
Signed-off-by: Yaniv Kaul <ykaul@redhat.com>