xlator_t: remove inconsistent alignment attribute
The 'stats' array inside 'xlator_t' was explicitly aligned to 128 bytes
(CAA_CACHE_LINE_SIZE = 128). Since this is a structure definition, it
caused the whole structure to be considered aligned to 128. That's not
problematic by itself as long as we really allocate the memory blocks
for 'xlator_t' objects at addresses multiple of 128 bytes, because the
compiler may use this fact to generate aligned access instructions
which are faster than unaligned ones.

Since we are not really allocating 'xlator_t' objects at that alignment,
the assumption that the compiler does is wrong, causing a crash.

Fixes: #3417
Change-Id: I33b5021035aae24f0b5ea14e255873da79fc0625
Signed-off-by: Xavi Hernandez <xhernandez@redhat.com>