fix gluster volume status timeout
Once one of nodes' network is suddenly dropped, it
takes several seconds for glusterd to perceive this change.

However, timeout may occurs during this period, when
other nodes treat the dropped one as normal, send
some request to it, and wait for response.

Meawhile, above behavior may also extend the update time
of the dropped node, which in turn causes timeouts to
occur more frequently during that period.

By considering this, a pre-check is seem to be feasible to make
sure the latest status of each node can be always obtained in time.

Fixes: gluster#3443

Change-Id: I1d3bed47f9bec0e8395efb7769ed4d85be26f61c
Signed-off-by: ChenJinhao <chen.jinhao@zte.com.cn>