tests: new testing tool
A new tool gftest has been created to automate and simplify the
preparation of an environment to run tests. It can also process
the full set of regression tests in parallel to significantly
reduce the total execution time.

It's written in python and requires these components:

- libvirt
- qemu-kvm
- gnupg
- python-click

All other modules should already be present in a default installation.

Using this tool is quite simple:

1. Create the VM

```
gftest create [OPTIONS] <name>

Options:
  --template <name> Template to use (default: centos7)
  --port <num>      Port to access the VM (default: 2222)
  --key <path>      Private key for SSH (default: ~/.ssh/id_rsa)
  --cpus <num>      Number of virtual cores of the VM
  --memory <num>    Memory assigned to the VM in GiB
  --disk <num>      Disk space for the VM in GiB
```

This creates a KVM VM named <name> that can be accessed through the
specified port on the local host (i.e. `ssh -p <port> root@127.0.0.1`)
using the specified private key. All other option will be taken from
the template if not present.

This process also creates a container inside the VM with all the
required packages and configurations to build Gluster and run the
tests.

2. Build the code

```
gftest build [OPTIONS] <name>

Options:
  --commit <id>  SHA/name of the commit to compile
```

This copies the current git repo to the VM and compiles the specified
commit in a container image.

3. Start the workers

```
gftest spawn [OPTIONS] <name>

Options:
  --workers <num>  Number of workers to create
  --space <num>    Space allocated for running tests in GiB
```

This starts the specified number of container instances of the compiled
Gluster image.

4. Run the tests

```
gftest run <name> <output dir>
```

This runs the full set of tests in parallel in all available containers
and puts the results into the <output dir>.

It also collects statistics for each test and CPU and memory state
during the full run.

5. Access the VM or container for manual testing/debugging

```
gftest sh <name> [<idx>]
```

Without `idx` it opens a shell session to the VM. With an index it opens
a shell session to the idx-th testing container.

There are some additional commands:

- `gftest kill <name>`
  Stops and destroys all testing containers.

- `gftest shutdown <name>`
  Gracefully shuts down the VM.

- `gftest poweroff <name>`
  Forcibly stops the VM.

- `gftest poweron <name>`
  Starts the VM.

Updates: #3469
Change-Id: I169877a4c5197d001bf2822ad7116559f7d78754
Signed-off-by: Xavi Hernandez <xhernandez@redhat.com>