stack.h: cleanups
Mainly cleanups	- removal of dead code,	re-order code to access	variables in order they	are defined, etc.
Importantly, assignment	of THIS	or old_THIS earlier in the macros, so they can be re-used again	in the logs.
Also - removal of redundant strerror() from the	logs - it's already done when we pass errno.

Removal of extra calls to 'THIS' and lastly, reorder of variables in the structures to reduce padding.

(This is all in preparation to move to the new logging facility in the future)

Updates: #1000
Signed-off-by: Yaniv Kaul <ykaul@redhat.com>