glusterd, cli: prefer binary uuid to avoid extra conversions
Prefer binary representation of UUID in peer, friend and snapshot
management code, which is expected to simplify the things and avoid
extra UUID conversions between text and binary representations
within glusterd. This comes with a presumably less critical overhead
of doing a few more conversions on client side, which is needed for
printing different objects' status in both textual and XML form.

Signed-off-by: Dmitry Antipov <dantipov@cloudlinux.com>
Updates: #1000