tests: improve security (#3548)
* Do not use a fixed password of 'pass' for nroot

Instead, we use a random enough password, which would be sufficient
for the minute this user is open. There should be at least around 32 000
differents passwords, which mean testing 533 passwords on the live server per
second. I suspect our builders aren't fast enough for that.

* Use a secure location for the script

Since the original script is in /tmp, a regular non root user could create
the file with lax permissions. The call to 'cat' would work, but then the
attacker could simply replace the content fast enough, and it would
be called each time ssh is run, which is quite often given the nature
of the georep feature testing.

This patch fix that by using mktemp to create a safe directory that
can't be accessed by a non root user, thus preventing the attack.

Change-Id: I92ef7a3c1cc8e57e9ffea8669fb21883efb81831
Signed-off-by: Michael Scherer <misc@redhat.com>