posix/gfid2path: Brick crashed if parent gfid link is not available
If the parent gfid link is unavailable for some reason, then the
brick will be crashed since it fails to find the path during gfid
to path conversion. Because there are no error whether the readlink
failed or not.

There is a race condition when resolving GFID split brain where the
parent dir could have renamed to a landfill location, so there can
be a case where parent gfid link is not available while a gfid to path
request in progress.

Change-Id: Icd71ccdb4ea2f9e454246c490e65cb609768ea02
Signed-off-by: Mohammed Rafi KC <rafi.kavungal@iternity.com>