Multiple files: reduce redirection for measure_latency variable
As far as I can see, we have the ctx variable in the call_stack_t structure only to be able to access the measure_latency boolean.
Example from create_frame():

    if (frame->root->ctx->measure_latency) {
        timespec_now(&stack->tv);
        memcpy(&frame->begin, &stack->tv, sizeof(stack->tv));
    }

I did not see any other use for the ctx variable apart from that boolean.
We can instead just copy that boolean and reduce one level of indirection (which seem to be happening in create_frame(), FRAME_DESTROY() and STACK_UNWIND_STRICT macro)

Updates: #1874
Signed-off-by: Yaniv Kaul <ykaul@redhat.com>