configure.ac: Fix AC_PROG_LEX warning
With autoconf 2.70, the following warnings are emitted:

configure.ac:544: warning: AC_PROG_LEX without either yywrap or noyywrap
is obsolete
./lib/autoconf/programs.m4:716: _AC_PROG_LEX is expanded from...
./lib/autoconf/programs.m4:709: AC_PROG_LEX is expanded from...
configure.ac:544: the top level

Autoconf docs[1] now states that usage of AC_PROG_LEX without choosing
one of the yywrap or noyywrap options is deprecated. It is usually
better to use noyywrap and define the yywrap function yourself, as this
almost always renders the LEXLIB unnecessary.

The behaviour of the argument on autoconf < 2.70 is to ignore the
argument, so there are no issues with adding the option.

[1]
https://www.gnu.org/savannah-checkouts/gnu/autoconf/manual/autoconf-
2.71/autoconf.html#index-AC_005fPROG_005fLEX-1

Signed-off-by: Anoop C S <anoopcs@cryptolab.net>